Add 'file' and 'rnm' to Synocli-File
- redesign cross/ncursesw to make independent of cross/ncurses
- create links for specific hashes of rhash
- add 'file' and 'rnm' to diyspk
- remove links on uninstall only if still linking to this package
- add creation of c++ library to cross/gmp

+++ spk/synocli-file/src/service-setup.sh
@@ -1,6 +1,6 @@
-
-COMMANDS="jdupes less lessecho lesskey tree ed2k-link edonr256-hash edonr512-hash ghost-hash has160-hash magnet-link"
-COMMANDS+="rhash sfv-hash tiger-hash tth-hash whirlpool-hash nano rnano ncdu"
+COMMANDS="jdupes less lessecho lesskey tree rhash rnm nano rnano ncdu file"
+RHASH_COMMAND_LINKS="ed2k-link edonr256-hash edonr512-hash gost-hash has160-hash magnet-link sfv-hash tiger-hash tth-hash whirlpool-hash"
+MC_COMMAND_LINKS="mc mcedit mcdiff mcview"
 
 service_postinst ()
 {
@@ -10,26 +10,54 @@ service_postinst ()
             ln -s "${SYNOPKG_PKGDEST}/bin/$cmd" "/usr/local/bin/$cmd"
         fi
     done
+    
+    # Set custom links for rhash
+    if [ -e "${SYNOPKG_PKGDEST}/bin/rhash" ]; then
+        for cmd in $RHASH_COMMAND_LINKS
+        do
+            ln -s "${SYNOPKG_PKGDEST}/bin/rhash" "/usr/local/bin/$cmd"
+        done
+    fi
 
     # Set custom links for midnight commander
-    ln -s ${SYNOPKG_PKGDEST}/bin/mc-utf8 /usr/local/bin/mc
-    ln -s ${SYNOPKG_PKGDEST}/bin/mc-utf8 /usr/local/bin/mcedit
-    ln -s ${SYNOPKG_PKGDEST}/bin/mc-utf8 /usr/local/bin/mcdiff
-    ln -s ${SYNOPKG_PKGDEST}/bin/mc-utf8 /usr/local/bin/mcview
+    if [ -e "${SYNOPKG_PKGDEST}/bin/mc-utf8" ]; then
+        for cmd in $MC_COMMAND_LINKS
+        do
+            ln -s "${SYNOPKG_PKGDEST}/bin/mc-utf8" "/usr/local/bin/$cmd"
+        done
+    fi
 }
 
 service_postuninst ()
 {
+    # remove links created by this package only
+    
     for cmd in $COMMANDS
-        do
+    do
         if [ -e "/usr/local/bin/$cmd" ]; then
-            rm -f "/usr/local/bin/$cmd"
+            if [ "$(readlink /usr/local/bin/$cmd)" == "${SYNOPKG_PKGDEST}/bin/$cmd" ]; then
+                rm -f "/usr/local/bin/$cmd"
+            fi
+        fi
+    done
+
+    # Remove custom links for rhash
+    for cmd in $RHASH_COMMAND_LINKS
+    do
+        if [ -e "/usr/local/bin/$cmd" ]; then
+            if [ "$(readlink /usr/local/bin/$cmd)" == "${SYNOPKG_PKGDEST}/bin/rhash" ]; then
+                rm -f "/usr/local/bin/$cmd"
+            fi
         fi
     done
 
     # Remove custom links for midnight commander
-    rm -f /usr/local/bin/mc
-    rm -f /usr/local/bin/mcedit
-    rm -f /usr/local/bin/mcdiff
-    rm -f /usr/local/bin/mcview
+    for cmd in $MC_COMMAND_LINKS
+    do
+        if [ -e "/usr/local/bin/$cmd" ]; then
+            if [ "$(readlink /usr/local/bin/$cmd)" == "${SYNOPKG_PKGDEST}/bin/mc-utf8" ]; then
+                rm -f "/usr/local/bin/$cmd"
+            fi
+        fi
+    done
 }
